% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeCRT_package.R
\docType{package}
\name{geeCRT}
\alias{geeCRT}
\title{geeCRT: a package for implementing the bias-corrected generalized estimating equations in analyzing cluster randomized trials}
\description{
geeCRT: a package for implementing the bias-corrected generalized estimating equations in analyzing cluster randomized trials
}
\section{geeCRT functions}{

The simbinPROBIT function performs correlated binary outcome data simulation using the multivariate probit method
The simbinCLF function performs correlated binary outcome data simulation using the conditional linear family method
The cpgeeSWD function performs cluster-period generalized estimating equations for estimating the marginal mean and correlation parameters in cross-sectional stepped wedge cluster randomized trials
The geemaee function performs matrix-adjusted generalized estimating equations on estimating the marginal mean and correlation parameters in cluster randomized trials
}

