% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cpgeeSWD.R
\name{print.cpgeeSWD}
\alias{print.cpgeeSWD}
\title{The print format for cpgeeSWD output}
\usage{
\method{print}{cpgeeSWD}(x, ...)
}
\arguments{
\item{x}{The object of cpgeeSWD output}

\item{...}{further arguments passed to or from other methods}
}
\value{
The output from \code{\link{print}}
}
\description{
The print format for cpgeeSWD output
}
